Move self-serve dashboard to /app/settings and add portal config
