+++ site/.env.example
@@ -37,3 +37,4 @@ RUNPOD_VISION_ENDPOINT_ID="replace-with-endpoint-id"
 STRIPE_SECRET_KEY="replace-with-stripe-secret-key"
 STRIPE_WEBHOOK_SECRET="replace-with-stripe-webhook-signing-secret"
 STRIPE_VISION_PRICE_ID="replace-with-stripe-price-id"
+STRIPE_PORTAL_CONFIGURATION_ID="replace-with-stripe-portal-configuration-id"
